requirements.txt
redmail
python-dotenv